requirements.txt
# Phase 1: Ingestion & Database
trafilatura>=1.6.0
youtube-transcript-api>=0.6.0
pydantic>=2.0.0
pytest>=7.0.0

# Phase 2: Concept Extraction & Canonicalization
nltk>=3.8.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Phase 3: Concept Graph (DAG)
faiss-cpu>=1.7.0
networkx>=3.0
scipy>=1.10.0
